+++ lib/src/parameterized_group_base.dart
@@ -60,5 +60,6 @@ void parameterizedGroup(
       onPlatform: onPlatform,
       retry: retry,
     ),
+    null,
   ).executeGroup(body);
 }